requirements.txt
sqlparse==0.2.4
django==2.2
djangorestframework==3.9.2
django-phone-field
djongo==1.2.38

django-filter